Refactor api methods to clean up their purposes a bit
