+++ extension/twitch/api/authorised-request.js
@@ -1,0 +1,9 @@
+const createApiRequest = require('./request')
+const context = require('../../context')
+
+const createAuthorisedApiRequest = params =>
+  createApiRequest(
+    Object.assign({}, params, { auth: context.twitch.auth })
+  )
+
+module.exports = createAuthorisedApiRequest
